+++ README.md
@@ -6,10 +6,10 @@
 
 This is a base project for the Software Architecture course in 2023/2024. It is a basic application composed of several components.
 
-- User service. Express service that handles the insertion of new users in the system.
-- Auth service. Express service that handles the authentication of users.
-- Gateway service. Express service that is exposed to the public and serves as a proxy to the two previous ones.
-- Webapp. React web application that uses the gateway service to allow basic login and new user features.
+- **User service**. Express service that handles the insertion of new users in the system.
+- **Auth service**. Express service that handles the authentication of users.
+- **Gateway service**. Express service that is exposed to the public and serves as a proxy to the two previous ones.
+- **Webapp**. React web application that uses the gateway service to allow basic login and new user features.
 
 Both the user and auth service share a Mongo database that is accessed with mongoose.
