+++ authservice/auth-service.js
@@ -2,7 +2,7 @@ const express = require('express');
 const mongoose = require('mongoose');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
-const User = require('./model')
+const User = require('./auth-model')
 
 const app = express();
 const port = 8002;
