+++ authservice/auth-service.test.js
@@ -1,10 +1,16 @@
 const request = require('supertest');
 const { MongoMemoryServer } = require('mongodb-memory-server');
 const bcrypt = require('bcrypt');
-const User = require('./model');
+const User = require('./auth-model');
 
 let mongoServer;
 let app;
+
+//test user
+const user = {
+  username: 'testuser',
+  password: 'testpassword',
+};
 
 async function addUser(user){
   const hashedPassword = await bcrypt.hash(user.password, 10);
@@ -21,6 +27,8 @@ beforeAll(async () => {
   const mongoUri = mongoServer.getUri();
   process.env.MONGODB_URI = mongoUri;
   app = require('./auth-service'); 
+  //Load database with initial conditions
+  await addUser(user);
 });
 
 afterAll(async () => {
@@ -30,12 +38,6 @@ afterAll(async () => {
 
 describe('Auth Service', () => {
   it('Should perform a login operation /login', async () => {
-    const user = {
-      username: 'testuser',
-      password: 'testpassword',
-    };
-    await addUser(user);
-
     const response = await request(app).post('/login').send(user);
     expect(response.status).toBe(200);
     expect(response.body).toHaveProperty('username', 'testuser');
